mt_site/requirements.txt
Django==3.1
psycopg2-binary==2.8.5
djangorestframework==3.11.0

pymodm==0.4.3
requests==2.24.0

Celery==4.4.6
elasticsearch==7.9.1
django-cors-headers==3.5.0
gunicorn==20.0.4
redis==3.5.3
django-celery-beat==2.1.0

dj-database-url==0.5.0
django-cache-url==3.2.2

django-redis==4.12.1
environs==9.3.1
